Add gitache_require_packages - for use from a submodule.
A local gitache directory no longer has to be a git clone,
it can be part of the super project. This change was made so
that OpenAI's Codex can work on gitache, as it can't work with
git submodules(!?)

This also adds the documentation of what I am working on,
and will soon be added to gitache-core: support for submodules
(or subdirectories - though not tested) to request their own
gitache packages from the super project.